social-syrup/phase4_pipeline/requirements.txt
anthropic==0.40.0
replicate==0.34.4
requests==2.32.3
requests-oauthlib==2.0.0
python-dotenv==1.0.1
Pillow==10.4.0
